first idea for getAll implemented

+++ helpers/generateShortURL.go
@@ -8,7 +8,7 @@ import (
 var alphabet = []byte("abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789")
 var alphabetLen = len(alphabet)
 
-func GenerateShortURL(urlLen int) string {
+func GenerateShortURL(urlLen uint8) string {
 	sb := strings.Builder{}
 
 	for range urlLen {
